Install working
The install just checks that the template and backup directories are ok and that the main file is exectutable. Then it checks if the program is aliased in popular config files such as `.bashrc` and `.zshenv` and if not asks the user if they would like the alias added. 

The install can be ran subsequent times to report if the program is setup properly and whether it is aliased and from where.